requirements.txt
requests==2.27.1
PyJWT==2.3.0
pandas==1.4.2